Redesigning the backend interface of NagVis.
Decided to throw the old backend code away for redesigning the backend interface
completely. Instead of querying the backend many times for each object the
objects which want to have some information from the backend add an entry to a
centrally managed backend queue.

Progressing the queue is started at some point were all objects added their
entries to the queue. With this situation it is possible to reduce the amount
of backend and datasource queries.

With the new backend interface the detail informations for an object like the
current plugin output are only fetched when they are really shown to the user.
When the current state of an object is only needed for calculating the summary
state of another object the backends will only be told to fetch the state
counts of that object.

At the moment the new interface has only been implemented for the MKLivestatus
backend. I will rework the ndomy backend too. Maybe the ndo2fs backend gets
dropped during this progress since MKLivestatus is a briliant alternative to
ndo2fs and ndo2fs is not widely used.