Finished recoding the backend interface
The recode has been completed for the ndomy and the mklivestatus backend. All problems should be solved now.

The backend functions are now called with a bunch of objects of the same type which have the same options (e.g. same only_hard_state setting).
This new structure decreases the number of queries to the data source and reduces the overhead needed for initiating the connection and so on.