Recoded map object references
All map objects need to have an object id which needs to be unique on
each map. All map objects which have no object id defined will get one
generated. This makes migration easier.

This new object id will be used on relative coords and during json
requests.

The automap is broken for the moment but will be fixed soon.