Using php redirect to for indexes instead of mod_rewrite
Using the php redirect seems to be better in some cases where https/http servers
are mixed. For example in OMD setups where using apache own mode and https in the
frontend and http in the backend apache servers.

Disabling this redirect by default in the hope that the php direct works better.