New feature: On demand maps (no config file needed)
Implemented "on demand" maps. Maps which only need a basic configuration
of global parameters like automaps and geomaps can now be configured and
created on demand by URL parameters. This mode is reached by simply
calling the map URL without "show=" parameter and a bunch of parameters
for the global section of the map instead.