FIX: Local code execution issue using modified .mo files
The php-gettext library in version 1.0.9 which was used to make NagVis
localizable on platforms which has no gettext support built into PHP.

A local user with privileges to create/modify the .mo files used by
NagVis could modify a file in a way that it executes custom code with
the privileges of the web server which is executing the NagVis code.

To more than 99% of NagVis installations (installed with the install.sh
installer or OMD) this is no special threat because a user which is able
to modify .mo files is able to modify the .php program files of NagVis. An
attacker would rather modify the .php files directly instead of injecting
code through the .mo files.

However, to fix this for most installations minor issue, we included gettext
in version 1.0.12 with this release which fixes the issue.