Introduce CorePDOHandler for authn/authz.
Combine the SQLite and MySQL drivers for authentication and
authorization into a single one that uses PHP's PDO database
abstraction layer.  Move the text of almost all the SQL queries
out of the functions and into a common array for easier tracking of
the differences needed between the database engines.  Make some
queries a bit more portable, largely moving from MySQL to PostgreSQL
syntax, so that they may be executed with no modification on other
database engines, too.