Correct size of single-line text input in popups
Prior to this commit single-line text inputs in popups would slightly
overflow. The border of the input field would not be properly visible.

With this commit this is improved slightly, and the ensuring we
correctly see the end of the text input.

In the case the input text is longer the then width of the text input,
one would need to "scroll" in the text field. This behaviour is unchanged
by this change.

Signed-off-by: Jacob Hansen <jhansen@op5.com>