Javascript error on refresh after updating an object
Currently, making a refresh right after modifying an object doesn't hide the context displayed. Context hide function were called to hide the whole dom_obj but was not hidding the element inside.

This commit will hide the context created upon right click and if it exist, it will make sure that it will be hidden after a mouse event. Also added a check on the condition for handling mouse clicks to avoid an error prompt, since it's expected to be null after the mouse event.

Signed-off-by: Eunice Remoquillo <eremoquillo@itrsgroup.com>