FIX: Disable state preview for worldmaps in Checkmk snapin
The state previews do not work correctly for worldmaps. To be able to
calculate the state preview, we would need the area of the map the user
would see when opening the map. To be able to do so, we need this
information when rendering the snapin: a) the viewport size and b) the
NagVis javascript code. Both is not available in this situation, so we
can not correctly render it. In the past the computation was done based
on information cached in NagVis, which could produce aggregated states,
but these might be misleading as they did not necessarily represent the
state of the objects you see in the map.