Validate session id
PHP does auto url-decoding for cookies. So if the session_id contains
malicious characters this could lead to http injections. Though an
attacker would need access to the cookie secret in order to calculate
valid signatures. Also the things that can be injected is very limited
since the `:` is used for splitting the cookie and cannot be injected.

This is according to my analysis not exploitable therefore no CVE.

CMK-23257