Fix User enumeration and Cleanup CMK session cookie validation
We send the cookie to the CMK RestAPI in order to check for the 2FA
validation.
So for Pre 2.0 this will fail anyways since it has no RestAPI. The
RestAPI will validate the cookie anyways so why do it twice? Also this
is problematic everytime Checkmk changes something with the cookie.

Now we only do some plausibility checks, mostly to avoid sending
newlines (which thanks to PHP are decoded automatically)

Previously when the user did not exist a different error message was
returned than if the session was invalid. This made it possible to
enumerate existing users.

We assigned CVE-2025-39665 with a CVSS score 6.9
(`CVSS:4.0/AV:N/AC:L/AT:N/PR:N/UI:N/VC:L/VI:N/VA:N/SC:N/SI:N/SA:N`)

CMK-26372